database: reorganize models into their own separate files
