+++ database/src/models/follow.rs
@@ -1,0 +1,10 @@
+use schema::follows;
+
+/// Represents a following relationship `[source user] -> [target user]`.
+#[derive(Identifiable, Queryable, Associations, PartialEq, Debug)]
+#[table_name = "follows"]
+pub struct Follow {
+    pub id: i64,
+    pub source_id: i64,
+    pub target_id: i64,
+}
