+++ database/src/models/mod.rs
@@ -1,0 +1,15 @@
+//! Database models.
+//!
+//! Note: do _not_ change the ordering of fields in model structs!
+//! The ordering must match that in the generated schema, which
+//! you can obtain with `diesel print-schema`.
+
+pub use self::account::{Account, NewAccount};
+pub use self::follow::Follow;
+pub use self::status::{NewStatus, Status};
+pub use self::user::{NewUser, User};
+
+mod account;
+mod follow;
+mod status;
+mod user;
